requirements.txt
fastapi
uvicorn
numpy-stl
trimesh
python-multipart
cadquery
pytransform3d
